@@ -4,8 +4,7 @@ import re
 from functools import partial
 from pathlib import Path
 from types import MappingProxyType
-from typing import TYPE_CHECKING, TypeVar, cast
-from warnings import warn
+from typing import TYPE_CHECKING, cast
 
 import h5py
 import numpy as np
@@ -24,6 +23,7 @@ from ..compat import (
     _from_fixed_length_strings,
 )
 from ..experimental import read_dispatched
+from ..utils import warn
 from .specs import read_elem, write_elem
 from .specs.registry import IOSpec, write_spec
 from .utils import (
@@ -41,8 +41,7 @@ if TYPE_CHECKING:
 
     from .._core.file_backing import AnnDataFileManager
     from .._core.raw import Raw
-
-T = TypeVar("T")
+    from .._types import StorageType
 
 
 @no_write_dataset_2d
@@ -261,7 +260,7 @@ def read_h5ad(
 
     with h5py.File(filename, "r") as f:
 
-        def callback(func, elem_name: str, elem, iospec):
+        def callback(read_func, elem_name: str, elem: StorageType, iospec: IOSpec):
             if iospec.encoding_type == "anndata" or elem_name.endswith("/"):
                 return AnnData(**{
                     # This is covering up backwards compat in the anndata initializer
@@ -279,7 +278,7 @@ def read_h5ad(
             elif elem_name in {"/obs", "/var"}:
                 # Backwards compat
                 return read_dataframe(elem)
-            return func(elem)
+            return read_func(elem)
 
         adata = read_dispatched(f, callback=callback)
 
@@ -322,7 +321,7 @@ def read_dataframe_legacy(dataset: h5py.Dataset) -> pd.DataFrame:
         f"{dataset.name!r} was written with a very old version of AnnData. "
         "Consider rewriting it."
     )
-    warn(msg, OldFormatWarning, stacklevel=2)
+    warn(msg, OldFormatWarning)
     df = pd.DataFrame(
         _decode_structured_array(
             _from_fixed_length_strings(dataset[()]), dtype=dataset.dtype
